+++ apps/store/src/components/CartPage/CartPageProps.types.ts
@@ -1,0 +1,5 @@
+type ProductData = { name: string; cost: number; currency: string }
+
+export type CartPageProps = {
+  products: Array<ProductData>
+}
